+++ docs/src/components/ComponentDoc/ComponentSourceManager/ComponentSourceManager.ts
@@ -4,9 +4,10 @@ import * as React from 'react'
 
 import { ExampleSource } from 'docs/src/types'
 import { componentAPIs as APIdefinitions, ComponentAPIs } from './componentAPIs'
-import getExampleSource from './getExampeSource'
+import getExampleModule from './getExampeSource'
 
 export type ComponentSourceManagerRenderProps = ComponentSourceManagerState & {
+  component: React.ElementType
   handleCodeAPIChange: (newApi: keyof ComponentAPIs) => void
   handleCodeChange: (newCode: string) => void
   handleCodeFormat: () => void
@@ -22,6 +23,7 @@ export type ComponentSourceManagerProps = {
 }
 
 type ComponentSourceManagerAPIs = ComponentAPIs<{
+  component: React.ElementType
   sourceCode: ExampleSource | undefined
   supported: boolean
 }>
@@ -48,12 +50,13 @@ export default class ComponentSourceManager extends React.Component<
     super(props)
 
     const componentAPIs = _.mapValues(APIdefinitions, (definition, name: keyof ComponentAPIs) => {
-      const sourceCode = getExampleSource(props.examplePath, name)
+      const module = getExampleModule(props.examplePath, name)
 
       return {
         ...definition,
-        sourceCode,
-        supported: !!sourceCode,
+        component: module && module.component,
+        sourceCode: module ? module.source : '',
+        supported: !!module,
       }
     }) as ComponentSourceManagerAPIs
 
@@ -136,6 +139,7 @@ export default class ComponentSourceManager extends React.Component<
   render() {
     return this.props.children({
       ...this.state,
+      component: this.state.componentAPIs[this.state.currentCodeAPI].component,
       handleCodeAPIChange: this.handleCodeAPIChange,
       handleCodeChange: this.handleCodeChange,
       handleCodeFormat: this.handleCodeFormat,