+++ docs/src/examples/components/MenuButton/index.tsx
@@ -2,11 +2,13 @@ import * as React from 'react'
 
 import Rtl from './Rtl'
 import State from './State'
+import Usage from './Usage'
 
 const MenuButtonExamples = () => (
   <>
     <State />
     <Rtl />
+    <Usage />
   </>
 )